# docs/swagger.yaml
definitions:
  domain.WxGenealogy:
    properties:
      cover:
        type: string
      createAt:
        type: string
      createBy:
        type: integer
      id:
        type: integer
      title:
        type: string
      updateAt:
        type: string
      updateBy:
        type: integer
    type: object
  domain.WxGenealogyMember:
    properties:
      addressDetail:
        type: string
      addressID:
        type: integer
      age:
        type: integer
      alive:
        type: integer
      avatar:
        type: string
      birthday:
        type: string
      deadTime:
        type: string
      genealogyID:
        type: integer
      id:
        type: integer
      name:
        type: string
      parentID:
        type: integer
      phone:
        type: integer
      role:
        type: integer
      sex:
        type: integer
      userID:
        type: integer
    type: object
  genealogy.AssembleRsp:
    properties:
      genealogies:
        items:
          $ref: '#/definitions/genealogy.wxGenealogyAssemble'
        type: array
      total:
        type: integer
    type: object
  genealogy.wxGenealogyAssemble:
    properties:
      avatar:
        type: string
      cover:
        type: string
      createBy:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  main.GreeterResponse:
    properties:
      message:
        type: string
    type: object
  response.Result:
    properties:
      Code:
        type: integer
      Data: {}
      Msg:
        type: string
    type: object
  valobj.GenealogyMembersPagingReq:
    properties:
      genealogyId:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
    type: object
  valobj.GenealogyMembersPagingRsp:
    properties:
      genealogyMembers:
        items:
          $ref: '#/definitions/domain.WxGenealogyMember'
        type: array
      total:
        type: integer
    type: object
  valobj.GenealogyPagingReq:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      title:
        type: string
    type: object
  valobj.MyGenealogyReq:
    properties:
      userId:
        type: integer
    type: object
info:
  contact: {}
paths:
  /myapp/api/v1/genealogy/assemble:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/valobj.GenealogyPagingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Result'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/genealogy.AssembleRsp'
                  type: array
              type: object
      summary: 分页聚合查询家谱数据
      tags:
      - Genealogy
  /myapp/api/v1/genealogy/member/save:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.WxGenealogyMember'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
      summary: 保存家谱成员
      tags:
      - Genealogy
  /myapp/api/v1/genealogy/members:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/valobj.GenealogyMembersPagingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Result'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/valobj.GenealogyMembersPagingRsp'
                  type: array
              type: object
      summary: 分页查询家谱人员信息
      tags:
      - Genealogy
  /myapp/api/v1/genealogy/my:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/valobj.MyGenealogyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Result'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/domain.WxGenealogy'
                  type: array
              type: object
      summary: 查询自己的所有族谱
      tags:
      - Genealogy
  /myapp/api/v1/genealogy/save:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.WxGenealogy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
      summary: 保存家谱
      tags:
      - Genealogy
  /v1/greeter:
    get:
      operationId: "1"
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GreeterResponse'
      summary: Greeter
      tags:
      - Hello
swagger: "2.0"